Craig McLaughlin's PhD dissertation and other minor fixes.
This commit adds Craig McLaughlin's PhD dissertation to the "Theses
and Dissertations" section.

Also, this commit inserts the often erroneously omitted "The" in the
name "The University of Edinburgh".

Furthermore, for consistency every "PhD thesis" is changed to "PhD
dissertation".

There is one more thing: updated the reference for JFP paper "Effect
Handlers via Generalised Continuations".